Solved stopping movement issue + faster movement
When a key was released, the code switched the dx/dy variable to 0. This is fine for stopping when the key is released, but releasing the key while holding the key in the opposite direction stops that opposite movement for a bit, "lagging" the player. Boolean values were added that check to see if the player is moving in the opposite direction or not. If they are, then the code doesn't set dx/dy to 0 and instead keeps it as is. 
Also made the player faster. I'm thinking of having a "focus" option similar to Touhou games where the player moves slower as to allow for more precise movement