Added a shooting mechanic
The player now has the option to shoot a projectile. Limitations to this are that the player can only shoot once before the bullet goes off screen. Issue with multiple bullets is that multiple objects of the PlayerBullet class must be made, and differentiating between them would be hard. Maybe I can add a variable in the PlayerBullet class named "number" or something to that effect and that's how I can tell them apart. It wouldn't count to infinity, but it would cycle through maybe 10 or so (those being the max allowed on screen).