Increased the limit of bullets, added enemies, and bullet-enemy collision
The number of allowable bullets onscreen was increased to 20, making it so that the player can spam space to their hearts content
Enemies have been added to the game, and they're represented by triangles. The color of these triangles will represent their difficulty, going from green (easy) to hard (red)
Bullets can collide with enemies now. This causes the bullet to disappear, but it's intended to make the enemy despawn as well. Will work on it later